pbs-171--refactor: user sign api
